requirement.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
qrcode
django-cors-headers
channels
channels_redis
django-redis